🔧 New API moved to api/ dir
