+++ common/src/main/java/infinituum/fastconfigapi/api/config/FastConfigs.java
@@ -1,4 +1,4 @@
-package infinituum.fastconfigapi.apiV2.config;
+package infinituum.fastconfigapi.api.config;
 
 import java.util.HashMap;
 import java.util.Map;
