+++ fabric/src/main/java/infinituum/fastconfigapi/fabric/utils/ModScanner.java
@@ -1,9 +1,9 @@
 package infinituum.fastconfigapi.fabric.utils;
 
 import infinituum.fastconfigapi.PlatformHelper;
-import infinituum.fastconfigapi.apiV2.annotations.FastConfig;
-import infinituum.fastconfigapi.apiV2.config.FastConfigFile;
-import infinituum.fastconfigapi.apiV2.serializers.SerializerWrapper;
+import infinituum.fastconfigapi.api.annotations.FastConfig;
+import infinituum.fastconfigapi.api.config.FastConfigFile;
+import infinituum.fastconfigapi.api.serializers.SerializerWrapper;
 import net.fabricmc.loader.api.FabricLoader;
 import net.fabricmc.loader.api.ModContainer;
 import net.minecraft.util.Tuple;
@@ -16,7 +16,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
-import static infinituum.fastconfigapi.apiV2.config.FastConfigs.MOD_ID;
+import static infinituum.fastconfigapi.api.config.FastConfigs.MOD_ID;
 
 public class ModScanner {
     public static <T> Map<Class<T>, FastConfigFile<T>> getSidedConfigs(FastConfig.Side side) {
